refactor: extract restore test steps in a shared util
